requirements.txt
hugchat==0.4.19
huggingface-hub==0.30.2
torch==2.5.1
transformers==4.51.3
accelerate==1.2.1
python-dotenv==1.0.1